Cleaning up the build.
git-svn-id: http://svn.fortytwo.net/projects@121 4bf1a796-0a03-0410-9b82-c499c1e613c7

+++ troika/phase2/trunk/src/api/p2_compiler.c
@@ -372,10 +372,10 @@ printf( "---c gc 1---\n" ); fflush( stdout );
     size_after = p2_memory_manager__size( m );
 printf( "---c gc 2---\n" ); fflush( stdout );
 
-    printf( "Collected %i of %i objects (%i%%).",
+    printf( "Collected %i of %i objects (%.3g%%).",
         size_before - size_after,
         size_before,
-        ( ( size_before - size_after ) * 100 ) / size_before );
+        ( ( size_before - size_after ) * 100 ) / ( double ) size_before );
 }
